refactor: move cast-send to own file
