+++ cli/src/cmd/cast/mod.rs
@@ -10,4 +10,5 @@ pub mod find_block;
 pub mod interface;
 pub mod rpc;
 pub mod run;
+pub mod send;
 pub mod wallet;
